fix-issue1758: create audit plan failed
details: create audit plan failed due to audit plan name too long

direct reason: The length of the jwt token exceeds the upper limit
of varchar(255) defined in the data table, resulting in insert failure.
deep reason: the length of apn is too long,resulting in the generated
 jwt token being too long.
solution: Use MD5 to hash the name and apn that may be too long to
make the name and apn for generating each token become fixed-length,
thereby controlling the length of the token.

@linxiaotao