fix-issue1756: cancel using goroutine
method im.BatchCancelApprove() will UPDATE dingtalk_instance, and method BatchUpdateWorkflowStatus() will UPDATE workflow, which execute after the former, should wait for it.
if we use go im.BatchCancelApprove(), we cannot see the order in which the two functions are executed.
so we cancel the using of goroutine on method im.BatchCancelApprove()

@linxiaotao