feat(configuration): Added SSH public key acquisition and generation functions
Added API interfaces for acquiring SSH public keys and generating SSH key pairs, which are used to manage the SSH key configuration of the system. The GetSSHPublicKey interface is used to obtain the stored SSH public key, and the GenSSHKey interface is used to generate a new SSH key pair and save it in a global variable.