feat(utils): Added the ability to generate SSH key pairs
Added `GenerateSSHKeyPair` function to generate a 4096-bit RSA private key and the corresponding SSH public key. The private key is encoded in PEM format, and the public key format is the same as that generated by the `ssh-keygen` tool. This function will be used in scenarios where an SSH key pair is required.